dom/tests/mochitest/bugs/test_bug876098.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=876098
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 876098</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 876098 **/
  var div = document.createElement("div");
  // count has to be large enough to trigger ion-compilation
  var count = 2000;
  // Separate function to make sure nothing weird we do block the ion-compile
  (function() {
    for (var i = 0; i < count; ++i) {
      var span = document.createElement("span");
      span.x = "foo";
      div.appendChild(span);
    }
  })();

  SpecialPowers.gc();

  function allHaveProp() {
    var kids = div.childNodes;
    for (var i = 0; i < count; ++i) {
      if (kids[i].x != "foo") {
        return false;
      }
    }
    return true;
  }

  ok(allHaveProp(), "All spans should have the property we added");


  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=876098">Mozilla Bug 876098</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
